Create SidebarHeader component. Move outer and inner Sidebars to SidebarHeader component.
